requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-core==0.3.28
langchain-groq==0.2.2
langchain-text-splitters==0.3.4
langsmith==0.2.7
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.1
orjson==3.10.13
packaging==24.2
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
